requirements.txt
# Core ML and Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.2

# Deep Learning
torch==2.0.1
torchvision==0.15.2
transformers==4.33.2
tensorflow==2.13.0

# Graph Neural Networks
torch-geometric==2.3.1
dgl==1.1.2

# Explainable AI
shap==0.42.1
lime==0.2.0.1

# API and Web Framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
requests==2.31.0

# Database and Caching
redis==4.6.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.19

# Streaming and Message Queue
kafka-python==2.0.2
pyspark==3.4.1

# Feature Store
feast==0.32.0

# Monitoring and Observability
prometheus-client==0.17.1
grafana-api==1.0.3
elasticsearch==8.8.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.0.3
loguru==0.7.0
click==8.1.7

# Security
cryptography==41.0.3
python-jose==3.3.0
passlib==1.7.4

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Time Series
prophet==1.1.4
statsmodels==0.14.0

# NLP
nltk==3.8.1
spacy==3.6.1
textblob==0.17.1

# Configuration Management
hydra-core==1.3.2
omegaconf==2.3.0

# Development Tools
black==23.7.0
flake8==6.0.0
mypy==1.5.1
pre-commit==3.3.3
